requirements.txt
streamlit
feedparser
google-generativeai
openai
beautifulsoup4
duckduckgo-search
json_repair
gTTS
requests